[250807] 1.axios 전역 코드 추가 2.공지사항 메인 /리스트 컴포넌트 분리
