+++ src/components/Layout.jsx
@@ -1,7 +1,7 @@
 import NavBar from "./NavBar"; // 사이드바 컴포넌트
 
 import { BrowserRouter, Outlet } from "react-router-dom";
-import SupportMain from "../pages/SupportMain.jsx";
+import Support from "../pages/Support";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 import DailyLedger from "../pages/DailyLedger";
@@ -16,15 +16,13 @@ export default function Layout() {
 
         {/* 메인 콘텐츠 영역 */}
         <main className="flex flex-1 bg-white overflow-y-auto overflow-x-hidden justify-center pt-10 pb-10 flex-wrap">
-
           <Routes>
-            <Route path="/support/*" element={<SupportMain />} />
+            <Route path="/support/*" element={<Support />} />
           </Routes>
 
           <Outlet />
         </main>
       </div>
     </BrowserRouter>
-
   );
 }
