+++ src/features/board/support/SuportMain.js
@@ -1,0 +1,53 @@
+import SupportList from "./SupportList";
+export default function SupportMain() {
+  return (
+    <div>
+      <section className="flex flex-col gap-6 mb-10">
+        <h1 className="text-xl text-center">고객센터</h1>
+        <div className="flex flex-col lg:flex-row gap-6">
+          <div className="card basis-1/2 custom-card1 rounded-box h-32 grow gap-3 justify-center p-10 shadow-md">
+            <p>
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 24 24"
+                strokeWidth={1.5}
+                stroke="currentColor"
+                className="size-6 inline mr-1"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  d="M2.25 6.75c0 8.284 6.716 15 15 15h2.25a2.25 2.25 0 0 0 2.25-2.25v-1.372c0-.516-.351-.966-.852-1.091l-4.423-1.106c-.44-.11-.902.055-1.173.417l-.97 1.293c-.282.376-.769.542-1.21.38a12.035 12.035 0 0 1-7.143-7.143c-.162-.441.004-.928.38-1.21l1.293-.97c.363-.271.527-.734.417-1.173L6.963 3.102a1.125 1.125 0 0 0-1.091-.852H4.5A2.25 2.25 0 0 0 2.25 4.5v2.25Z"
+                />
+              </svg>
+              02-000-0000
+            </p>
+
+            <a href="mailto:savee@savee.com">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 24 24"
+                strokeWidth={1.5}
+                stroke="currentColor"
+                className="size-6 inline mr-1"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  d="M21.75 6.75v10.5a2.25 2.25 0 0 1-2.25 2.25h-15a2.25 2.25 0 0 1-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0 0 19.5 4.5h-15a2.25 2.25 0 0 0-2.25 2.25m19.5 0v.243a2.25 2.25 0 0 1-1.07 1.916l-7.5 4.615a2.25 2.25 0 0 1-2.36 0L3.32 8.91a2.25 2.25 0 0 1-1.07-1.916V6.75"
+                />
+              </svg>
+              savee@savee.com
+            </a>
+          </div>
+          <div className="card basis-1/2 cursor-pointer custom-card2 border rounded-box grid h-32 grow place-items-center shadow-md">
+            <p>Qna</p>
+          </div>
+        </div>
+      </section>
+      <SupportList></SupportList>
+    </div>
+  );
+}
