--- src/pages/SupportAdd.jsx
@@ -1,4 +1,0 @@
-import SupportInput from "../features/board/support/SupportInput";
-export default function SupportAddPage() {
-  return <SupportInput></SupportInput>;
-}
